Make multi-module
* Break spray marshalling into `mwundo-spray` module
* Rename primary artifact to `mwundo-core` to avoid upgrade confusion when spray disappears
* Bump minor version
* Update README with new module names and versions
