+++ core/src/main/scala/com/monsanto/labs/mwundo/JTSGeometryImplicits.scala
