+++ core/src/test/scala/com/monsanto/labs/mwundo/GeometryClipper_UT.scala
