+++ spray/src/test/scala/com/monsanto/labs/mwundo/GeoJsonFormatPropTest.scala
@@ -1,7 +1,7 @@
-package com.monsanto.labs.mwundo.spray
+package com.monsanto.labs.mwundo
 
 import com.monsanto.labs.mwundo.GeoJson._
-import com.monsanto.labs.mwundo.spray.GeoJsonFormats._
+import com.monsanto.labs.mwundo.GeoJsonFormats._
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalacheck.Gen._
 import org.scalacheck.Prop._
@@ -12,7 +12,7 @@ import spray.json._
 
 class GeoJsonFormatPropTest extends FunSpec with Checkers with ParallelTestExecution {
 
-  implicit override val generatorDrivenConfig = PropertyCheckConfiguration(workers = 4)
+  implicit override val generatorDrivenConfig: PropertyCheckConfiguration = PropertyCheckConfiguration(workers = 4)
   import GeoJsonGenerators._
 
   private def marshalAndUnmarshal[T](t: T)(implicit jf: JsonFormat[T]) = {
